Fixed couple bugs
And did some performance testing and seeing 50-60 FPS in Java, runs great except for the black map coming into view problem. In Cheerpj environment, this version runs but at a slower FPS. In the GraphicsThread the value "Thread.sleep(10);" needs changed to 30 to work in Cheerpj on my slow dual core computer.