fix-client-freeze-on-mob-death
Fix client freeze on mob death

This commit fixes a bug that caused the client to freeze when a mob was killed. The bug was caused by a race condition where the client would sometimes not receive the final HP update for a mob before it was removed from the world.

This commit fixes the bug by:

- Adding a new method to `DuskEngine.java` to broadcast HP updates to all players in an area.
- Updating `Battle.java` to use the new HP update method and to send a final HP update when a mob is killed.
- Updating `Dusk.java` to not draw the HP bar for dead entities and to handle HP updates for entities that have already been removed.